Task2/requirements.txt
pandas==1.5.3 
matplotlib==3.7.1 
seaborn==0.12.2 
scikit-learn==1.2.2 
numpy==1.24.3 
jupyter==1.0.0 
python-pptx==0.6.21